+++ packages/web-components/src/components/style-picker/__stories__/_story.defs.ts
@@ -1,0 +1,25 @@
+/**
+ * @license
+ *
+ * Copyright IBM Corp. 2023
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import { POPOVER_ALIGNMENT } from '@carbon/web-components/es/components/popover/defs.js';
+
+export const alignOptions = [
+  POPOVER_ALIGNMENT.TOP,
+  POPOVER_ALIGNMENT.TOP_LEFT,
+  POPOVER_ALIGNMENT.TOP_RIGHT,
+  POPOVER_ALIGNMENT.BOTTOM,
+  POPOVER_ALIGNMENT.BOTTOM_LEFT,
+  POPOVER_ALIGNMENT.BOTTOM_RIGHT,
+  POPOVER_ALIGNMENT.LEFT,
+  POPOVER_ALIGNMENT.LEFT_BOTTOM,
+  POPOVER_ALIGNMENT.LEFT_TOP,
+  POPOVER_ALIGNMENT.RIGHT,
+  POPOVER_ALIGNMENT.RIGHT_BOTTOM,
+  POPOVER_ALIGNMENT.RIGHT_TOP,
+];